fix(arch): force stopped state when ok
There are situations when the fdCount is already zero and the
eventsource wasn't explicitly stopped yet (wasn't in a "STOPPING"
state). So when the last FD was shut down it would not transition into
the "STOPPED" state.
This commit explicitly checks when the shutdown is called if we already
have no FDs and then transition into the STOPPED state here.